requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.0
fastapi==0.103.0
h11==0.16.0
idna==3.10
kafka-python==2.2.7
psycopg2-binary==2.9.7
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dotenv==1.0.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.13.2
uvicorn==0.23.2
